move children to a different property
It looks like there would need to be a lot of changes to accomodate the
union, so going to try putting children in its own property for now.

We can make the changes, but it's just going to take some work. Maybe
it'll be cleaner to just make the value in every relationship a list.