Modify the BaseGPTIndex delete function to call docstore delete along with other deletes.
The reason is that there is a bug where the docstore metadata will not delete the doc_id
when calling delete. This code change has been verified and fixes the issue.
raise_error is set to False, because the doc_id might not be in "docstore/data",
but it would be in "docstore/metadata".