Adding support for encoding argument on SimpleDirectoryReader
This allows users to pass on the encoding they want to use to open
the files in a given directory.
Having this would remove the need to convert files into a specific
encoding to get them to work and also avoid potentially unoticed
data loss by loading documents with the wrong encoding.