feat: add int VID for NebulaGraph & parse comment in schema
pr: #7402

This helps those non-KG graph storage to be enabled
by GraphRAG and Text2Cypher, where the vid could be
numeric and even typed as int64.

Previously it assumed we are running a KG that vertex
ID was always string.

Add comment fields in Schema to be parsed, this helped
the case when schema field name are short,
but comment with rich information in text2cypher