llms/openai: fix Azure OpenAI streaming
The Azure OpenAI API, when used with streaming, may return a response
with just the content filtering information but no content in 'choices'.
Instead of returning an empty ChatResponse (which would affect other
users like StreamingAgentChatResponse), just skip that first response
(there is nothing we do with that data after all).