use json.loads instead of ast
At the moment ast.eval is used to decode a JSON string. However, if a tool uses booleans, ast fails because in the JSON format the boolean keywords are lowercase. 

To fix this, we switch to json.loads as primary decoder and we use ast as fallback option when json.loads failes.