Fix handling of empy queries
Sometimes the retriever will get an empty query. For example, the
AutoRetriever may decide to query only the metadata. Before this
commit, that caused an exception due to trying to get an embedding
on an empty string. This commit fixes things so that the embedding
step is skipped when the query string is empty.