Only capture valid python function names in ReAct output
Some LLMs output extra information at the end of the function
name and cause llama_index to fail to find the function they
want to call. Tighten the regular expression to only accept
valid python function names (as used in the lookup) and
ignore additional output.

This makes more LLMs work with the ReAct agent, even if
they do not follow the instructions exactly.