Implements a Node Parser using embeddings for Semantic Splitting
Provided an implementation of BaseEmbedding, allows Documents to be split into nodes based on semantic similarity. First,
the text is split into sentences using the provided sentence splitter.  Then, sentences are recursively grouped together
to form a list of all possible continguous sentence permutations in the document.

Embeddings are then generated for each individual sentence and each group of sentences, then the cosine distance is calculated
to determine the similarity between each group of sentence and the following one. Percentiles are then calculated for each
similarity to normalize the distribution and identify outliers.

Finally, each similarity percentile is evaluated against a user-provided threshold to determine breakpoints for semantic groups
of sentences.  Providing a higher breakpoint percentile will result in larger nodes, and visa versa.  The default value of 95th
percentile seems to be a good default in most scenarios.