Updated OpenInferenceCallbackHandler to deal with ChatMessage responses
When handling the LLM events in the on_event_end method, if there is a
response payload object, this may be of type ChatResponse, which has a
string representation with format "<role>: <message>". However we want
only the message part. This commit takes this into account, and if a
ChatResponse objects is received instead of getting the string
representation the code gets message string.