Make Google Drive Reader serializable (#12286)
Allow passing in client config, authorized user info, and service account keys
directly in addition to supporting file names. Store the file contents in the
object so they can be written to disk. Also allow passing a field to indicate
the reader is running in the cloud, in which case the tokens should not be
written to disk.