Improve efficiency of MongoDB reader
MongoDB reader was not using projections.
So, in case, someone is fetching a small set of fields then the full document would still be read in those cases.

This commit fixes this by adding a projection consisting of all the `field_names` and `metadata_names`. This should reduce the bandwidth usage and improve the performance of MongoDB reader, especially, when reading a few fields from big documents.