Update dynamic_llm.py -> Made output parsers functional again
I tried to make the output parsers for the dynamic llm (both with and without properties) as flexible and as robust to inconsistences in LLM output as possible. With weaker LLMs, sometimes the LLM used double quotations mixed with single quotations, and sometimes just single quotations. These weren't properly handled by the very rigid previous output parsers which resulted in empty knowledge graphs despite the LLMs returning an output full of triplets with minor format inconsistencies. 

This should serve as a temporary fix for now, it's better than what was initially here but not perfect as we never know what the LLM might come up with next ^^'

The ideal would be to use PyDantic to define a proper rigid output but I didn't have the time to do it for now.