Fix incorrect type in OpenAI token usage report
Corrects an incorrect type assumption when extracting OpenAI token
usage reported in the response from (a)client.chat.completions.create.

The response is actually a ChatCompletion or ChatCompletionChunk, which
are pydantic objects [1]. The _get_response_token_counts assumed that
the response is a dict, with a fallback of not returning anything if the
response was anything other than a dict. Because of the fallback, the
incorrect type assumption didn't result in an error, but did result in
missing token counts. The extracted token counts are returned as
additional_kwargs in the ChatResponse.

[1] https://github.com/openai/openai-python/blob/3d780f260a1d26ab612dde0f870db114a42273bd/src/openai/types/chat/chat_completion.py#L76

Closes: #15523