fix: Fix GraphRAG notebook V2
Simplified regex for getting entities, removed expectation of the following characters "(", ")", "{", "}"

These characters are no longer a part of the node.text this is why the regex wasn't returning anything. With these changes the GraphRAG works.
Why these characters are no longer a part of the node.text I don't know, if these characters should be a part of the text of the node this will not resolve the issue.