Don't manually set return_properties when querying Weaviate
Returning all properties is the standard in the Weaviate v4 client. Manually setting this to a list of the attribute names is not required and also does not work when non-primitive types are present in the properties (e.g. a list of nested objects in one of the properties): https://github.com/weaviate/weaviate-python-client/issues/630#issuecomment-1924152015