Make Weaviate Vector Store integration work with complex properties (#17129)
* Don't manually set return_properties when querying Weaviate

Returning all properties is the standard in the Weaviate v4 client. Manually setting this to a list of the attribute names is not required and also does not work when non-primitive types are present in the properties (e.g. a list of nested objects in one of the properties): https://github.com/weaviate/weaviate-python-client/issues/630#issuecomment-1924152015

* Add a test for reading data from a collection that includes a complex property

* Bump version of llama-index-vector-stores-weaviate to 1.2.3

* Get rid of commented out code

* Only assert core of what test should check, add comment

* Simplify test_can_query_collection_with_complex_property_types()

* Simplify test_can_query_collection_with_complex_property_types(), do not try to close client if the variable was not successfully set

---------

Co-authored-by: Logan <logan.markewich@live.com>