feat: support typing.Annotation for adding parameter descriptions
Using pydantic.Field works fine, but typing.Annotation is a more Pythonic and less library-dependent way to
add description strings to function parameters. So now you can write a function like this:

    def tmp_function(x: Annotated[int, 'An integer'] = 3) -> str:
        return str(x)

And it will pass the description of x in the tool description to the LLM. It doesn't remove the existing behavior,
but it gives a new option that makes use of Python's built-in way to do this, rather than relying on Pydantic.