Support file extensions in GoogleDriveReader
Previously GoogleDrives folders that had file extentions besides Google Sheets/Docs/Presentation
would have their extensions stripped because the filename was the Google Drive file ID which
doesn't have an extension. However this meant that reading other types of file formats (like PDFs)
wouldn't work and the "file_extractor" capability wouldn't be used, instead everything would just be reading bytes.
That works ok for .txt or .md docs, but for a PDF it's generates a bunch of garbage.

This change adds the file extension if it can be returned, otherwise it'll just add empty string and shouldn't create any issues

It's possible in GoogleDrive to also remove the extension, but that should then fall back to the prior behavior

This could be a breaking change if some files were previously being parsed as bytes/text but would now be getting parsed by the file_extractor