llamaindex-cli to handle glob patterns correctly
This change modifies the llamaindex-cli such that it can
handle the --files argument to properly handle glob patterns.

In order to handle globs like the example given in the issue,
the number of arguments (nargs) value must be set to + in order
to return a list of files or patterns in this case.

Because argparse now returns a list, so restructuring on how
the files are processed needed to be changed along with the
signature of the function.

Fixes #11798

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>