Ref #17877 - Fixing Aquery with Postgres with IVFFLAT
As described in the issue, vector store index querying was not successful
when the PGVectorStore was initialized with the `ivfflat_probes` argument.
This problem is potentially due to the behavior of psycopg3 sending parameters
separately, using [server-side binding](https://www.psycopg.org/psycopg3/docs/basic/from_pg2.html#server-side-binding),
which may mis-behave when performed asynchronously. However, this is just
a hypothesis and may be related to a different issue.

The proposed solution works successfully and formats the variable directly
in the `SET` string.