Ensure that Github reader uses timeout and retries params
Previously, GithubRepositoryReader would ignore the retries parameter,
and in _generate_documents it would also ignore the timeout parameter.
This commit ensures that retries is passed to all calls to the
GithubClient.

Some users may have their code broken by this fix if they have passed a
subclass of BaseGithubClient whose `get_*` methods don't support the
`timeout` and/or `retries` keywords. The default implementation
`GithubClient` does support these keywords. The existing code is already
somewhat broken because we're already passing `timeout` in some places.
I'd guess only a few users will be affected by this, but just in case
I'm putting it in the next minor version.