watsonxembeddings: Make persistent conn configurable
In the IBM watsonx Python SDK, persistent connections were enabled in
version 1.1.2 and higher. However, this causes problems in some
environments where network connections are cut from time to time. This
creates a connection error and the inference fails.

Add a `persistent_connection` boolean for the WatsonxEmbeddings class set to
True by default (to match upstream), but allow it to be disabled if
needed.

Signed-off-by: Major Hayden <major@mhtx.net>