Strip code block backticks from QueryFusionRetriever llm response (#18825)
We've observed cases where the resonse for the QueryFusionRetriever
from gemini-1.5-pro-001 will be wrapped with a code block. So instead of:

"first query\nsecond query"

it returns:

"```\nfirst query\nsecond query\n```"

So instead of parsing ["first query", "second query"],
we get ["```", "first query"] as the two derived queries.
Obviously, the "```" query ends up retreiving poor quality
nodes.

This may not be a widespread issue, but I thought this
change was small and safe enough to be worth implementing
to guard against this problem.

(I did not add a unit test because I do not see existing
tests for the QueryFusionRetriever).