chore: update the embeddings mixedbread version
Calling the mixedbread embedding currently yields a "503 Service
Temporary Unavailable". The current version of the mixedbread client
used is deprecated and not compatible with the api anymore.

Let's use the new "mixedbread" pypi package which is the supported
Mixedbread python client. This client is mostly compatible with the
old one, except that the TruncationStrategy got removed.

For backwards-compatibility the llama-index-integration will remain
named mixedbreadai.